+++ RELEASE_NOTES.md
@@ -1,4 +1,7 @@
 # v0.0.81
 
 -   fixed excessive allowance bug
+-   now properly estimates correct gas limit, and adds 10% tolerance
+-   set maxPriorityFeePerGas to 30 gwei on polygon to avoid stuck transactions
 -   refactored some code & general code quality improvements
+-   fix for goerli tokens in getLinkDetails